package :: relocate APIs location
