+++ app/src/main/java/com/teamnk/kimiljung/api/APIProvider.kt
@@ -1,5 +1,8 @@
 package com.teamnk.kimiljung.api
 
+import com.teamnk.kimiljung.feature.login.LoginAPI
+import com.teamnk.kimiljung.feature.register.RegisterAPI
+
 val loginAPIProvider: LoginAPI by lazy {
     RetrofitClient.getRetrofit()!!.create(LoginAPI::class.java)
 }
